Renamed load_reporting.{c,h} to load_reporting_plugin.{c,h}
