+++ binding.gyp
@@ -892,8 +892,8 @@
         'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_fallback.c',
         'src/core/ext/filters/client_channel/resolver/dns/native/dns_resolver.c',
         'src/core/ext/filters/client_channel/resolver/sockaddr/sockaddr_resolver.c',
-        'src/core/ext/filters/load_reporting/load_reporting.c',
         'src/core/ext/filters/load_reporting/load_reporting_filter.c',
+        'src/core/ext/filters/load_reporting/load_reporting_plugin.c',
         'src/core/ext/census/base_resources.c',
         'src/core/ext/census/context.c',
         'src/core/ext/census/gen/census.pb.c',
