+++ gRPC-Core.podspec
@@ -440,8 +440,8 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/client_channel/resolver/fake/fake_resolver.h',
                       'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_ev_driver.h',
                       'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper.h',
-                      'src/core/ext/filters/load_reporting/load_reporting.h',
                       'src/core/ext/filters/load_reporting/load_reporting_filter.h',
+                      'src/core/ext/filters/load_reporting/load_reporting_plugin.h',
                       'src/core/ext/census/aggregation.h',
                       'src/core/ext/census/base_resources.h',
                       'src/core/ext/census/census_interface.h',
@@ -697,8 +697,8 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper_fallback.c',
                       'src/core/ext/filters/client_channel/resolver/dns/native/dns_resolver.c',
                       'src/core/ext/filters/client_channel/resolver/sockaddr/sockaddr_resolver.c',
-                      'src/core/ext/filters/load_reporting/load_reporting.c',
                       'src/core/ext/filters/load_reporting/load_reporting_filter.c',
+                      'src/core/ext/filters/load_reporting/load_reporting_plugin.c',
                       'src/core/ext/census/base_resources.c',
                       'src/core/ext/census/context.c',
                       'src/core/ext/census/gen/census.pb.c',
@@ -934,8 +934,8 @@ Pod::Spec.new do |s|
                               'src/core/ext/filters/client_channel/resolver/fake/fake_resolver.h',
                               'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_ev_driver.h',
                               'src/core/ext/filters/client_channel/resolver/dns/c_ares/grpc_ares_wrapper.h',
-                              'src/core/ext/filters/load_reporting/load_reporting.h',
                               'src/core/ext/filters/load_reporting/load_reporting_filter.h',
+                              'src/core/ext/filters/load_reporting/load_reporting_plugin.h',
                               'src/core/ext/census/aggregation.h',
                               'src/core/ext/census/base_resources.h',
                               'src/core/ext/census/census_interface.h',
